Refactor code: Add delete functionality for food items
Added a new function in the foodHandler.go file to handle the deletion of food items. The function validates the token, deletes the specified food item from the database, and returns a success message. This functionality is accessed through the /deleteFood/:food_id route in the foodRoute.go file.

Refactor CSS files for improved layout and styling

Made changes to the global.css and meal.css files to improve the layout and styling of the web application. Removed unnecessary CSS properties, adjusted table width, and added new styles for displaying the total calories and quantity of foods in a meal.

Refactor code: Update middleware and CSS files

Updated the middleware and CSS files to enhance the functionality and appearance of the web application. Made changes to the foodHandler.go file to include a missing struct field. Also, made updates to the global.css file to improve the layout and styling.

Refactor code: Update pop-up IDs and remove unused files and functions

Updated the pop-up IDs in the btCreateFoodConfirm.mjs file to correctly retrieve the meal ID from the URL. Additionally, removed unused files and functions from the project to improve code cleanliness and maintainability.